+++ examples/kitchensink/login-with-single-sign-on.ts
@@ -6,7 +6,7 @@ const { ROOT_URL, NODE_ENV, UNCHAINED_CLOUD_ENDPOINT } = process.env;
 
 const loginWithSingleSignOn = async (remoteToken, unchainedAPI: Context) => {
   try {
-    const thisDomain = new URL(ROOT_URL).hostname;
+    const domain = process.env.UNCHAINED_COOKIE_DOMAIN || new URL(ROOT_URL).hostname || 'localhost';
     const result = await fetch(UNCHAINED_CLOUD_ENDPOINT, {
       method: 'POST',
       headers: {
@@ -26,10 +26,7 @@ const loginWithSingleSignOn = async (remoteToken, unchainedAPI: Context) => {
       throw new Error(json.errors[0].message);
     }
 
-    if (
-      thisDomain === 'localhost' ||
-      thisDomain === json?.data?.controlConsumeSingleSignOnToken?.domain
-    ) {
+    if (domain === 'localhost' || domain === json?.data?.controlConsumeSingleSignOnToken?.domain) {
       // create sso user if not exist and login
       const ssoUserId =
         (await unchainedAPI.modules.users.findUser({ username: 'sso' }))?._id ||
@@ -50,8 +47,8 @@ const loginWithSingleSignOn = async (remoteToken, unchainedAPI: Context) => {
         unchainedAPI,
       );
       const expires = new Date(tokenExpires || new Date().getTime() + 100000);
-      const authCookie = cookie.serialize('token', token, {
-        domain: thisDomain || 'localhost',
+      const authCookie = cookie.serialize(process.env.UNCHAINED_COOKIE_NAME, token, {
+        domain,
         httpOnly: true,
         expires,
         path: '/',