+++ packages/api/src/context.ts
@@ -16,9 +16,9 @@ export const setCurrentContextResolver = (newContext: UnchainedContextResolver) 
 export const createContextResolver =
   (unchainedAPI: UnchainedCore, roles: any, version: string): UnchainedContextResolver =>
   async ({ req, res, ...apolloContext }) => {
-    const loaders = await instantiateLoaders(req, unchainedAPI);
-    const userContext = await getUserContext(req, unchainedAPI);
-    const localeContext = await getLocaleContext(req, unchainedAPI);
+    const loaders = await instantiateLoaders(req, res, unchainedAPI);
+    const userContext = await getUserContext(req, res, unchainedAPI);
+    const localeContext = await getLocaleContext(req, res, unchainedAPI);
 
     return {
       req,